restructured pages/components for locations page
